refactor: declaration files (and tests); `connect(): Realm`
